+++ src/constants/post.ts
@@ -132,6 +132,9 @@ export const MAX_VILLAGERS = 10;
 export const MAX_CART_ITEMS = 100;
 export const MAX_ITEM_TEXT_LINES = MAX_CART_ITEMS;
 export const MIN_ITEM_TEXT_LINES = 2;
+// 붙여넣기 일괄 추가 — 라인별 검색 결과 수와 재확인 후보 상한
+export const HITS_PER_LINE = 5;
+export const MAX_REVIEW_CANDIDATES = 2;
 
 export const ARTWORKS_WITH_FAKES = [
 	'아카데믹한 명화',
