+++ src/hooks/item/query/useMatchCatalogItems.ts
@@ -1,26 +1,17 @@
 import { useMutation } from '@tanstack/react-query';
 
 import { searchClient } from '@/config/firebase';
-import { MAX_ITEM_TEXT_LINES } from '@/constants/post';
-import { CatalogItem } from '@/types/catalog';
+import { HITS_PER_LINE, MAX_ITEM_TEXT_LINES } from '@/constants/post';
 import {
-	buildCleanedSearchTerms,
-	BulkLineMatchResult,
-	classifyCatalogHits,
-} from '@/utilities/bulkItemMatching';
+	type BulkLineMatchResult,
+	type BulkMatchOutput,
+	type FailedLineMatchResult,
+	type FoundLineMatchResult,
+	type NeedsReviewLineMatchResult,
+} from '@/types/bulkItemMatching';
+import { CatalogItem } from '@/types/catalog';
+import { buildCleanedSearchTerms, classifyCatalogHits } from '@/utilities/bulkItemMatching';
 import { getTrimmedNonEmptyLines } from '@/utilities/itemTextLines';
-
-export type BulkMatchResult = BulkLineMatchResult;
-export type BulkFoundResult = Extract<BulkMatchResult, { status: 'found' }>;
-export type BulkNeedsReviewResult = Extract<BulkMatchResult, { status: 'needsReview' }>;
-export type BulkFailedResult = Extract<BulkMatchResult, { status: 'failed' }>;
-
-export interface BulkMatchOutput {
-	results: BulkMatchResult[];
-	foundResults: BulkFoundResult[];
-	needsReviewResults: BulkNeedsReviewResult[];
-	failedResults: BulkFailedResult[];
-}
 
 /**
  * 사용자가 붙여넣은 전체 텍스트를 검색할 줄 목록으로 정리
@@ -36,13 +27,11 @@ const parseSearchLines = (text: string): string[] => {
 	return lines;
 };
 
-const HITS_PER_LINE = 5;
-
 const searchCatalogLine = async (
 	line: string,
 	searchTerm: string,
 	source: 'original' | 'cleaned',
-): Promise<BulkMatchResult> => {
+): Promise<BulkLineMatchResult> => {
 	const { results } = await searchClient.searchForHits<CatalogItem>({
 		requests: [
 			{
@@ -62,10 +51,12 @@ const searchCatalogLine = async (
 	return classifyCatalogHits({ line, searchTerm, source, hits });
 };
 
-const matchLine = async (line: string): Promise<BulkMatchResult> => {
+const matchLine = async (line: string): Promise<BulkLineMatchResult> => {
+	// 원본 텍스트로 검색
 	const originalResult = await searchCatalogLine(line, line, 'original');
 	if (originalResult.status !== 'failed') return originalResult;
 
+	// 정제된 검색어로 재시도
 	for (const cleanedTerm of buildCleanedSearchTerms(line)) {
 		const cleanedResult = await searchCatalogLine(line, cleanedTerm, 'cleaned');
 		if (cleanedResult.status !== 'failed') return cleanedResult;
@@ -86,13 +77,13 @@ const matchCatalogItems = async (text: string): Promise<BulkMatchOutput> => {
 	return {
 		results: lineResults,
 		foundResults: lineResults.filter(
-			(result): result is BulkFoundResult => result.status === 'found',
+			(result): result is FoundLineMatchResult => result.status === 'found',
 		),
 		needsReviewResults: lineResults.filter(
-			(result): result is BulkNeedsReviewResult => result.status === 'needsReview',
+			(result): result is NeedsReviewLineMatchResult => result.status === 'needsReview',
 		),
 		failedResults: lineResults.filter(
-			(result): result is BulkFailedResult => result.status === 'failed',
+			(result): result is FailedLineMatchResult => result.status === 'failed',
 		),
 	};
 };
